feat(integrity_checks): add dynamic lookup check
