+++ src/Helpers/PageHelpers/IPageHelper.cs
@@ -7,7 +7,6 @@ namespace form_builder.Helpers.PageHelpers
 {
     public interface IPageHelper
     {
-        void ValidateDynamicLookUpObject(List<Page> pages, string formName);
         Task<FormBuilderViewModel> GenerateHtml(Page page, Dictionary<string, dynamic> viewModel, FormSchema baseForm, string guid, FormAnswers formAnswers, List<object> results = null);
 
         void SaveAnswers(Dictionary<string, dynamic> viewModel, string guid, string form, IEnumerable<CustomFormFile> files, bool isPageValid, bool appendMultipleFileUploadParts = false);
