+++ src/Utils/ServiceCollectionExtensions/ServiceCollectionExtensions.cs
@@ -281,6 +281,7 @@ namespace form_builder.Utils.ServiceCollectionExtensions
             services.AddSingleton<IFormSchemaIntegrityCheck, CurrentEnvironmentSubmitSlugsCheck>();
             services.AddSingleton<IFormSchemaIntegrityCheck, DateValidationsCheck>();
             services.AddSingleton<IFormSchemaIntegrityCheck, DocumentDownloadCheck>();
+            services.AddSingleton<IFormSchemaIntegrityCheck, DynamicLookupCheck>();
             services.AddSingleton<IFormSchemaIntegrityCheck, EmailActionsCheck>();
             services.AddSingleton<IFormSchemaIntegrityCheck, EmptyBehaviourSlugsCheck>();
             services.AddSingleton<IFormSchemaIntegrityCheck, GeneratedIdConfigurationCheck>();
