+++ src/Validators/FormSchema/IntegrityChecks/CurrentEnvironmentSubmitSlugsCheck.cs
@@ -1,10 +1,8 @@
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using form_builder.Enum;
-using form_builder.Extensions;
 using form_builder.Models;
+using form_builder.Extensions;
 using Microsoft.AspNetCore.Hosting;
 
 namespace form_builder.Validators.IntegrityChecks
