+++ tests/unit-tests/UnitTests/Helpers/PageHelperTests.cs
@@ -100,38 +100,6 @@ namespace form_builder_tests.UnitTests.Helpers
                 _mockActionHelper.Object);
         }
 
-        [Theory]
-        [InlineData("local", "int", "Fake", "TestToken", "https://myapi.com")] // 1) No Match:  environment for current Environment Name
-        [InlineData("local", "", "Fake", "TestToken", "https://myapi.com")] // 2) No Environment Name
-        [InlineData("local", "local", "", "TestToken", "https://myapi.com")] // 3) No Provider Name
-        [InlineData("local", "local", "Test_Provider", "TestToken", "https://myapi.com")] // 4) No Provider found.. ( Just have Fake to select )
-        [InlineData("local", "local", "Fake", "TestToken", "")] // 5) No URL to hit
-        [InlineData("local", "local", "Fake", "", "https://myapi.com")] // 6) No Auth Token
-        [InlineData("int", "int", "Fake", "TestToken", "http://myapi.com")] // 7) No https if not on local
-        public void ValidateDynamicLookUpObject_ShouldThrowError_IfNotValidLookupProperties(string environmentName, string lookupEnv, string provider, string authToken, string url)
-        {
-            // Arrange
-            _mockHostingEnv.Setup(_ => _.EnvironmentName).Returns(environmentName);
-
-            var element = new ElementBuilder().WithType(EElementType.Radio).WithLookup("dynamic").Build();
-            element.Properties.LookupSources = new List<LookupSource>
-            {
-                new LookupSource
-                {
-                    EnvironmentName = lookupEnv,
-                    Provider = provider,
-                    AuthToken = authToken,
-                    URL = url
-                }
-            };
-
-            var page = new PageBuilder().WithElement(element).Build();
-            List<Page> pages = new() { page };
-
-            // Act & Assert
-            Assert.Throws<ApplicationException>(() => _pageHelper.ValidateDynamicLookUpObject(pages, "test-form"));
-        }
-
         [Fact]
         public async Task GenerateHtml_ShouldAddOptions_WhenFormContainsDynamicLookup()
         {
