+++ tests/unit-tests/UnitTests/Validators/IntegrityChecks/CurrentEnvironmentSubmitSlugsCheckTests.cs
@@ -50,7 +50,7 @@ namespace form_builder_tests.UnitTests.Validators.IntegrityChecks
 
             // Assert
             Assert.False(result.IsValid);
-            Assert.True(result.Messages.Any(_ => _.Equals($"FAILURE - No SubmitSlug found in form 'test-name' for environment 'local'.")));
+            Assert.Contains(result.Messages, _ => _.Equals($"FAILURE - No SubmitSlug found in form 'test-name' for environment 'local'."));
         }
     }
 }
